fix: advance payment due-date formula for companies (חברות)
Companies now correctly receive advance payment due dates on the 15th of
the period's own month, not the following month. Self-employed (עצמאי)
continues to use the next-month formula.

- Fetch full Business object in generator to inspect business_type
- Add branching logic: COMPANY → same month; OSEK_* → next month
- Import BusinessType enum

Fixes: Companies were systematically assigned due dates one month too late.

+++ app/vat_reports/services/data_entry_common.py
@@ -31,7 +31,7 @@ def recalculate_totals(
     work_item_repo: VatWorkItemRepository,
     invoice_repo: VatInvoiceRepository,
     item_id: int,
-) -> tuple[float, float]:
+) -> tuple[Decimal, Decimal]:
     """Recompute output / input VAT totals from stored invoices (single query)."""
     output_vat, input_vat = invoice_repo.sum_vat_both_types(item_id)
     output_net, input_net = invoice_repo.sum_net_both_types(item_id)
